pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "note-taking-app"
version = "0.1.0"
description = "AI voice note taking application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.0",
    "langchain>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
]

[tool.setuptools]
packages = {find = {}}

[tool.poetry]
package-mode = false
